Medtronic 0.8.1-SNAPSHOT - wear changes - refactoring of decoder - started work on bolus, tbr, resumes

+++ app/src/main/java/info/nightscout/androidaps/plugins/pump/medtronic/comm/ui/MedtronicUIPostprocessor.java
@@ -137,7 +137,7 @@ public class MedtronicUIPostprocessor {
         Duration dur = new Duration(clockDTO.localDeviceTime.toDateTime(DateTimeZone.UTC),
             clockDTO.pumpTime.toDateTime(DateTimeZone.UTC));
 
-        clockDTO.timeDifference = dur.getStandardSeconds();
+        clockDTO.timeDifference = (int)dur.getStandardSeconds();
 
         MedtronicUtil.setPumpTime(clockDTO);
