+++ app/src/main/java/info/nightscout/androidaps/plugins/pump/medtronic/data/dto/ClockDTO.java
@@ -12,5 +12,7 @@ public class ClockDTO {
 
     public LocalDateTime pumpTime;
 
-    public long timeDifference; // s
+    //public Duration timeDifference;
+
+    public int timeDifference; // s
 }
